+++ src/assets/data/thanks.ts
@@ -19,7 +19,7 @@ export const orderedThanksKeys: Array<ThanksKeys> = [
 
 /** 感谢名单 */
 export default {
-    lastUpdated: "2025-11-20",
+    lastUpdated: "2025-12-14",
     list: {
         starBuilders,
         starGuides,