+++ src/components/shared/CreationCard.vue
@@ -1,3 +1,5 @@
+<!-- 视频卡片 -->
+
 <template>
     <div class="creation-item">
         <div class="video-box">