+++ src/components/ui/LinksNav.vue
@@ -2,20 +2,28 @@
 
 <template>
     <nav id="links-nav">
+        <!-- 站内导航 -->
         <template v-for="link in internalLink" :key="link.name">
-            <RouterLink :to="link.path" active-class="on">
+            <RouterLink :to="link.path" active-class="on" :title="link.name">
                 {{ link.name }}
             </RouterLink>
         </template>
+
+        <!-- 站外导航 -->
+        <!-- <template v-for="item in publicExternalLinks" :key="item.title">
+            <a :href="item.href" :title="item.title" target="_blank" rel="noopener noreferrer">
+                {{ item.title }}
+            </a>
+        </template> -->
     </nav>
 </template>
 
 <script setup lang="ts">
 import { useNavLinks } from "@/utils";
-import { publicPages } from "@/router/routes/publicPage";
+import { publicPages, publicExternalLinks } from "@/router/routes";
 
 // 内链
-const internalLink = useNavLinks(publicPages)
+const internalLink = useNavLinks(publicPages);
 </script>
 
 <style scoped lang="less">
@@ -27,22 +35,22 @@ const internalLink = useNavLinks(publicPages)
     gap: 1.25rem;
     max-width: 37.5rem;
     text-align: center;
-    transition: all 0.3s ease-in-out;
     white-space: nowrap;
 
     a {
         padding: 0 0.25rem;
         color: @text-color;
         text-decoration: none;
+        border-bottom: 0.0625rem solid transparent;
 
         &:hover {
-            border-bottom: 0.0625rem solid @link-on-color;
+            border-bottom-color: @link-on-color;
         }
 
         &.on {
             color: @inline-title;
             font-weight: bold;
-            border-bottom: 0.0625rem solid @inline-title;
+            border-bottom-color: @inline-title;
 
             &:hover {
                 border-bottom-color: @inline-title;