+++ src/utils/index.ts
@@ -3,7 +3,7 @@ export * from "./factory/anti-shake";
 // export * from "./factory/dateUtils";
 export * from "./factory/device";
 export * from "./factory/nav";
-// export * from "./factory/url";
+export * from "./factory/url";
 export * from "./factory/useMediaQuery";
 export * from "./factory/VideoID";